+++ src/main/java/com/yohanii/lostandfound/domain/post/Post.java
@@ -1,6 +1,7 @@
 package com.yohanii.lostandfound.domain.post;
 
 import com.yohanii.lostandfound.domain.user.User;
+import com.yohanii.lostandfound.dto.post.PostEditRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -36,10 +37,19 @@ public class Post {
     private LocalDateTime updatedTime;
 
     @Builder
-    public Post(User user, String title, String content, PostType type) {
+    public Post(User user, String title, String content, PostType type, LocalDateTime createdTime, LocalDateTime updatedTime) {
         this.user = user;
         this.title = title;
         this.content = content;
         this.type = type;
+        this.createdTime = createdTime;
+        this.updatedTime = updatedTime;
+    }
+
+    public void updatePost(PostEditRequestDto dto) {
+        this.title = dto.getTitle();
+        this.content = dto.getContent();
+        this.type = dto.getType();
+        this.updatedTime = LocalDateTime.now();
     }
 }