+++ src/main/java/com/yohanii/lostandfound/dto/post/PostEditRequestDto.java
@@ -1,0 +1,18 @@
+package com.yohanii.lostandfound.dto.post;
+
+import com.yohanii.lostandfound.domain.post.Post;
+import com.yohanii.lostandfound.domain.post.PostType;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class PostEditRequestDto {
+
+    private String title;
+    private String content;
+    private PostType type;
+
+}