+++ src/main/java/com/yohanii/lostandfound/web/post/PostController.java
@@ -3,8 +3,10 @@ package com.yohanii.lostandfound.web.post;
 import com.yohanii.lostandfound.domain.post.Post;
 import com.yohanii.lostandfound.domain.post.PostRepository;
 import com.yohanii.lostandfound.domain.user.User;
+import com.yohanii.lostandfound.dto.post.PostEditRequestDto;
 import com.yohanii.lostandfound.dto.post.PostSaveRequestDto;
 import com.yohanii.lostandfound.web.SessionConst;
+import com.yohanii.lostandfound.web.argumentresolver.Login;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -35,19 +37,21 @@ public class PostController {
     }
 
     @GetMapping("/{postId}")
-    public String post(@PathVariable Long postId, Model model) {
+    public String post(@Login User loginUser, @PathVariable Long postId, Model model) {
         Post post = postRepository.findById(postId);
 
+        log.info("loginUser={}", loginUser);
+        model.addAttribute("user", loginUser);
         model.addAttribute("post", post);
         return "posts/post";
     }
 
-    @GetMapping("/add")
-    public String postCreate(@ModelAttribute PostSaveRequestDto post) {
+    @GetMapping("/add-form")
+    public String postSaveForm(@ModelAttribute PostSaveRequestDto post) {
         return "posts/addPostForm";
     }
 
-    @PostMapping("/add")
+    @PostMapping
     @Transactional
     public String postSave(@Validated @ModelAttribute PostSaveRequestDto dto, BindingResult bindingResult, HttpServletRequest request) {
         if (bindingResult.hasErrors()) {
@@ -57,6 +61,38 @@ public class PostController {
         HttpSession session = request.getSession();
         User loginUser = (User) session.getAttribute(SessionConst.LOGIN_USER);
         postRepository.save(dto.toEntity(loginUser));
+        return "redirect:/posts";
+    }
+
+    @GetMapping("/{postId}/edit-form")
+    public String postEditForm(@PathVariable("postId") Long postId, @ModelAttribute PostEditRequestDto dto, Model model) {
+
+        Post findPost = postRepository.findById(postId);
+        dto.setTitle(findPost.getTitle());
+        dto.setContent(findPost.getContent());
+        dto.setType(findPost.getType());
+
+        model.addAttribute("post", findPost);
+
+        return "posts/editPostForm";
+    }
+
+    @PatchMapping("/{postId}")
+    @Transactional
+    public String postEdit(@PathVariable("postId") Long postId, @ModelAttribute PostEditRequestDto dto) {
+
+        Post findPost = postRepository.findById(postId);
+        findPost.updatePost(dto);
+
+        return "redirect:/posts/{postId}";
+    }
+
+    @DeleteMapping("/{postId}")
+    @Transactional
+    public String postDelete(@PathVariable("postId") Long postId) {
+
+        postRepository.delete(postId);
+
         return "redirect:/posts";
     }
 }