+++ src/test/java/com/yohanii/lostandfound/domain/post/PostRepositoryTest.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest
 @Transactional
@@ -41,13 +42,11 @@ class PostRepositoryTest {
     @Test
     void findById() {
         Post post1 = Post.builder()
-                .user(null)
                 .title("hello")
                 .content("hello123")
                 .type(PostType.LOST)
                 .build();
         Post post2 = Post.builder()
-                .user(null)
                 .title("hello2")
                 .content("hello1234")
                 .type(PostType.FOUND)
@@ -65,13 +64,11 @@ class PostRepositoryTest {
     @Test
     void findAll() {
         Post post1 = Post.builder()
-                .user(null)
                 .title("hello")
                 .content("hello123")
                 .type(PostType.LOST)
                 .build();
         Post post2 = Post.builder()
-                .user(null)
                 .title("hello2")
                 .content("hello1234")
                 .type(PostType.FOUND)
@@ -81,6 +78,24 @@ class PostRepositoryTest {
 
         List<Post> result = postRepository.findAll();
         assertThat(result).contains(post1, post2);
-        assertThat(result.size()).isEqualTo(2);
     }
+
+    @Test
+    void delete() {
+        Post post1 = Post.builder()
+                .title("hello")
+                .content("hello123")
+                .type(PostType.LOST)
+                .build();
+        postRepository.save(post1);
+        List<Post> postsBefore = postRepository.findAll();
+
+        postRepository.delete(post1.getId());
+        List<Post> postsAfter = postRepository.findAll();
+
+        assertThat(postsAfter.size()).isEqualTo(postsBefore.size() - 1);
+        assertThat(postsBefore).contains(post1);
+        assertThat(postsAfter).doesNotContain(post1);
+    }
+
 }